Added feature like getAllAccount/DeleteAccount/Withdraw and deposit Money

+++ src/main/java/com/banking_application/banking_app/dto/AccountDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class AccountDto {
     private Long id;
     private String accountHolderName;
-    private double balance;
+    private Long balance;
 }
