+++ src/main/java/com/banking_application/banking_app/entity/Account.java
@@ -19,5 +19,5 @@ public class Account {
     private Long id;
     @Column(name = "account_holder_name")
     private String accountHolderName;
-    private double balance;
+    private Long balance;
 }
